refactor bigquery cache into cache and processor
